Output files for crossVmap
Lines 79-99 produce and output the information needed for latere input to pf.crossVmap.

group.data.frame should be a table that python can import, and so should the tree.

These objects (or their paths) can be input into run_crossVmap to obtain the cross-validated feature table and the cross-validated groups.